Moving performance tests into separate file and removing using.
