+++ samples/LibuvWithNonAllocatingFormatters/Program.cs
@@ -7,7 +7,6 @@ using System.Buffers;
 using System.Buffers.Text;
 using System.Linq;
 using System.Net.Libuv;
-using System.Text;
 using System.Text.Formatting;
 using System.Text.Utf8;
 using System.Threading.Tasks;
