+++ samples/LowAllocationWebServer/LowAllocationWebServerLibrary/HttpServer.cs
@@ -9,7 +9,6 @@ using System.Text.Formatting;
 using System.Text.Utf8;
 using System.Text.Http;
 using System.Threading.Tasks;
-using System.Text;
 using System.Threading;
 using System.Buffers.Text;
