+++ samples/LowAllocationWebServer/LowAllocationWebServerLibrary/RoutingTable.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Buffers.Text;
-using System.Text;
 using System.Text.Http;
 using System.Text.Utf8;
