+++ samples/System.IO.Pipelines.Samples/Framing/Codec.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Net;
-using System.Text;
 using System.Text.Formatting;
 using System.Threading.Tasks;
 using System.IO.Pipelines.Networking.Libuv;
