+++ samples/System.IO.Pipelines.Samples/SampleBase/RawHttpClientSampleBase.cs
@@ -4,7 +4,6 @@
 
 using System.Linq;
 using System.Threading.Tasks;
-using System.Text;
 using System.Text.Formatting;
 using System.Buffers.Text;
