+++ samples/System.IO.Pipelines.Samples/SampleBase/RawHttpServerSampleBase.cs
@@ -4,7 +4,6 @@
 
 using System.Buffers.Text;
 using System.Net;
-using System.Text;
 using System.Text.Formatting;
 using System.Threading.Tasks;
