+++ src/System.Binary/System/Binary/BufferReader_BE.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Runtime;
 using System.Runtime.CompilerServices;
 
 namespace System.Buffers
