+++ src/System.Buffers.Experimental/System/Buffers/BytesReader.cs
@@ -3,7 +3,6 @@
 
 using System.Buffers.Text;
 using System.Runtime.CompilerServices;
-using System.Text;
 
 namespace System.Buffers
 {
