+++ src/System.IO.Pipelines.Text.Primitives/ReadableBufferExtensions.cs
@@ -3,7 +3,6 @@
 
 using System.Buffers;
 using System.Runtime.CompilerServices;
-using System.Text;
 using System.Text.Parsing;
 using System.Text.Utf8;
