+++ src/System.Numerics.Tensors/System/Numerics/CompressedSparseTensor.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.Text;
+﻿using System.Diagnostics;
 
 namespace System.Numerics
 {
