+++ src/System.Text.Primitives/System/Text/Encoders/Ascii/Ascii_casing.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Buffers;
 
 namespace System.Buffers
 {
