+++ src/System.Text.Primitives/System/Text/Extensions.Integer.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System.Buffers;
 
 namespace System.Buffers.Text
 {
