+++ src/System.Text.Primitives/System/Text/Formatters/Custom.Integer.cs
@@ -3,7 +3,6 @@
 
 using System.Buffers.Text;
 using System.Diagnostics;
-using System.Text;
 
 namespace System.Buffers
 {
