+++ src/System.Text.Primitives/System/Text/Formatters/Utf16.Integer.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Runtime.CompilerServices;
-using System.Text;
 
 namespace System.Buffers
 {
