+++ src/System.Text.Primitives/System/Text/Parsers/PrimitiveParser_Boolean.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Buffers.Text;
-using System.Text;
 
 namespace System.Buffers
 {
