+++ src/System.Text.Primitives/System/Text/Parsers/PrimitiveParser_Decimal.cs
@@ -3,7 +3,6 @@
 
 
 using System.Buffers.Text;
-using System.Text;
 
 namespace System.Buffers
 {
