+++ tests/System.Text.Primitives.Tests/Encoding/AsciiCasingTests.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.Buffers;
 using System.Text;
 using Xunit;
