+++ tests/System.Text.Primitives.Tests/Parsing/PrimitiveParserGuidTests.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Buffers;
-using System.Buffers.Text;
 using System.Globalization;
 using Xunit;
